Bug Fixes/ImprovEMentsw
Fixed a bug preventing the broken ground to destroy itself when required

fixed a bug which if a earth construct is spawned before the enemies are spawned it would crash the game.
Also fixed issue where if the enemy it was chasing was destroyed it would break and remain afk.

+++ MyProject2/Source/MyProject2/Characters/XPShard.h
@@ -44,7 +44,7 @@ public:
 	void OnOverlapBegin(UPrimitiveComponent* OverlappedComp, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult);
 
 
-	uint16 XPGainAmount;
+	int XPGainAmount;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Shard Type")
 	ShardTypes m_ShardType;
